requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
aiohttp>=3.9.1

# Azure AI & Bot Framework
azure-ai-projects==1.0.0b11
azure-identity>=1.15.0
azure-ai-agents==1.1.0b2
azure-core>=1.30.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
botbuilder-core==4.17.0
botbuilder-schema==4.17.0
botframework-connector==4.17.0

# Document Processing & PDF Generation
reportlab>=4.0.8
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2

# Email & Communication
sendgrid>=6.11.0
requests>=2.31.0

# Data Processing
pandas>=2.1.4
numpy>=1.24.3
beautifulsoup4>=4.12.2

# Web Search & News
newsapi-python>=0.2.7
feedparser>=6.0.10

# Logging & Monitoring
structlog>=23.2.0
